budget-auto-shutdown/requirements.txt
google-cloud-run
google-api-core
google-auth
requests
protobuf
functions-framework
